Split ViewerButton in its own styles file since it's used more than once
